[Fix] Improved Indent Spacing Tests
Improved Indent Spacking test to check for more than 4 spaces on the leading edge of a line as it was previously done. Improved testing for continuation lines to include mathmatical operators and boolean comparators.

Corrected and added Unit tests to use appropriate space in the multi-line strings and to validate the continuation line conditions.